+++ AvaloniaProjectRelations/Browser/IHtmlViewerVMFactory.cs
@@ -1,7 +1,0 @@
-namespace AvaloniaProjectRelations.Browser
-{
-    internal interface IHtmlViewerVMFactory
-    {
-        HtmlViewerVM Create(string content);
-    }
-}