+++ AvaloniaProjectRelations/Commands/ICancelCommandFactory.cs
@@ -1,9 +1,0 @@
-using Topelab.Core.Avalonia.Base;
-using Topelab.Core.Avalonia.Command;
-
-namespace AvaloniaProjectRelations.Commands
-{
-    internal interface ICancelCommandFactory : ICommandFactory<BaseVM>
-    {
-    }
-}