+++ AvaloniaProjectRelations/Main/IMainWindowVMFactory.cs
@@ -1,7 +1,0 @@
-namespace AvaloniaProjectRelations.Main
-{
-    internal interface IMainWindowVMFactory
-    {
-        MainWindowVM Create();
-    }
-}