+++ AvaloniaProjectRelations/MainControl/IMainControlVMChangeListener.cs
@@ -1,9 +1,0 @@
-namespace AvaloniaProjectRelations.MainControl
-{
-    using CmdTools.Shared;
-    using Topelab.Core.Avalonia.Base;
-
-    internal interface IMainControlVMChangeListener : IBaseVMChangeListener<MainControlVM, ProjectOptions>
-    {
-    }
-}