+++ AvaloniaProjectRelations/MainControl/IMainControlVMFactory.cs
@@ -1,7 +1,0 @@
-namespace AvaloniaProjectRelations.MainControl
-{
-    internal interface IMainControlVMFactory
-    {
-        MainControlVM Create(string[] args);
-    }
-}