+++ AvaloniaProjectRelations/MainControl/IMainControlVMInitializer.cs
@@ -1,7 +1,0 @@
-namespace AvaloniaProjectRelations.MainControl
-{
-    internal interface IMainControlVMInitializer
-    {
-        void Initialize(MainControlVM vm, bool isFirstInitialization = false);
-    }
-}