+++ Contracts/IElementRelationsGetterFactory.cs
@@ -1,7 +1,0 @@
-namespace CmdTools.Contracts
-{
-    public interface IElementRelationsGetterFactory
-    {
-        IElementRelationsGetter Create(FinderType finderType);
-    }
-}