+++ RelationsShared/Services/IProjectsServiceFactory.cs
@@ -1,9 +1,0 @@
-namespace RelationsShared.Services
-{
-    using System.Collections.Generic;
-
-    public interface IProjectsServiceFactory
-    {
-        IProjectsService Create(string solutionPath, IEnumerable<string> projectPaths);
-    }
-}